+++ saw-core-what4/src/SAWCoreWhat4/What4.hs
@@ -45,16 +45,12 @@ module SAWCoreWhat4.What4
   , symExprToValue
   ) where
 
-
-
 import qualified Control.Arrow as A
 
 import Data.Bits
 import Data.IntMap (IntMap)
 import qualified Data.IntMap as IntMap
 import Data.IORef
-import Data.Kind (Type)
-import Data.List (genericTake,intercalate)
 import Data.Map (Map)
 import qualified Data.Map as Map
 import Data.Ratio ((%))
@@ -63,10 +59,9 @@ import qualified Data.Set as Set
 import qualified Data.Text as Text
 import Data.Vector (Vector)
 import qualified Data.Vector as V
-import qualified Data.BitVector.Sized as BV
 
 import Data.Traversable as T
-import Control.Monad ((<=<), foldM, unless)
+import Control.Monad (foldM)
 import Control.Monad.State as ST (MonadState(..), get, put, StateT(..), evalStateT, modify)
 import Control.Monad.Trans.Class (MonadTrans(..))
 import Numeric.Natural (Natural)
@@ -79,14 +74,14 @@ import SAWCore.SATQuery
 import SAWCore.SharedTerm
 import SAWCore.Simulator.Value
 import SAWCore.FiniteValue (FirstOrderType(..), FirstOrderValue(..))
-import SAWCore.Module (ModuleMap, ResolvedName(..), ctorName, dtCtors, lookupVarIndexInMap)
-import SAWCore.Name (Name(..), VarName(..), toAbsoluteName, toShortName, toQualName)
+import SAWCore.Module (ModuleMap)
+import SAWCore.Name (Name(..), VarName(..), toShortName)
 import SAWCore.Term.Functor (FieldName)
 
 -- what4
 import qualified What4.Expr.Builder as B
 import           What4.Expr.GroundEval
-import           What4.Interface(SymExpr,Pred,SymInteger, IsExpr, SymFnWrapper(..),
+import           What4.Interface(SymExpr,Pred,SymInteger, IsExpr,
                                  IsExprBuilder,IsSymExprBuilder, BoundVar)
 import qualified What4.Interface as W
 import           What4.BaseTypes
@@ -95,54 +90,15 @@ import           What4.SWord (SWord(..))
 
 -- parameterized-utils
 import qualified Data.Parameterized.Context as Ctx
-import Data.Parameterized.Map (MapF)
-import qualified Data.Parameterized.Map as MapF
-import Data.Parameterized.Context (Assignment)
 import Data.Parameterized.Some
 
 -- saw-core-what4
+import SAWCoreWhat4.Common
 import SAWCoreWhat4.PosNat
 import SAWCoreWhat4.FirstOrder
 import SAWCoreWhat4.Panic
 import SAWCoreWhat4.ReturnTrip
-
----------------------------------------------------------------------
--- empty datatype to index (open) type families
--- for this backend
-data What4 (sym :: Type)
-
--- | A What4 symbolic array where the domain and co-domain types do not appear
---   in the type
-data SArray sym where
-  SArray ::
-    W.IsExpr (W.SymExpr sym) =>
-    W.SymArray sym (Ctx.EmptyCtx Ctx.::> itp) etp ->
-    SArray sym
-
--- type abbreviations for uniform naming
-type SBool sym = Pred sym
-type SInt  sym = SymInteger sym
-
-type instance EvalM (What4 sym) = IO
-type instance VBool (What4 sym) = SBool sym
-type instance VWord (What4 sym) = SWord sym
-type instance VInt  (What4 sym) = SInt  sym
-type instance VArray (What4 sym) = SArray sym
-type instance Extra (What4 sym) = What4Extra sym
-
-type SValue sym = Value (What4 sym)
-type SPrim sym  = Prims.Prim (What4 sym)
-
--- Constraint
-type Sym sym = IsSymExprBuilder sym
-
----------------------------------------------------------------------
-
-data What4Extra sym =
-  SStream (Natural -> IO (SValue sym)) (IORef (Map Natural (SValue sym)))
-
-instance Show (What4Extra sym) where
-  show (SStream _ _) = "<SStream>"
+import SAWCoreWhat4.Uninterp
 
 ---------------------------------------------------------------------
 --
@@ -666,9 +622,6 @@ selectV sym merger maxValue valueFn vx =
       p <- SW.bvAtLE sym vx (toInteger j)
       merger p (impl j (y `setBit` j)) (impl j y) where j = i - 1
 
-instance Show (SArray sym) where
-  show (SArray arr) = show $ W.printSymExpr arr
-
 arrayConstant ::
   W.IsSymExprBuilder sym =>
   sym ->
@@ -902,338 +855,11 @@ w4SolveBasic sym sc addlPrims varMap ref unintSet t =
      Sim.evalSharedTerm cfg t
 
 
-----------------------------------------------------------------------
--- Uninterpreted function cache
-
-{-
-data SymFnWrapper sym :: Ctx.Ctx BaseType -> Type where
-  SymFnWrapper :: !(W.SymFn sym args ret) -> SymFnWrapper sym (args Ctx.::> ret)
--}
-type SymFnCache sym = Map W.SolverSymbol (MapF (Assignment BaseTypeRepr) (SymFnWrapper sym))
-
-lookupSymFn ::
-  W.SolverSymbol -> Assignment BaseTypeRepr args -> BaseTypeRepr ty ->
-  SymFnCache sym -> Maybe (W.SymFn sym args ty)
-lookupSymFn s args ty cache =
-  do m <- Map.lookup s cache
-     SymFnWrapper fn <- MapF.lookup (Ctx.extend args ty) m
-     return fn
-
-insertSymFn ::
-  W.SolverSymbol -> Assignment BaseTypeRepr args -> BaseTypeRepr ty ->
-  W.SymFn sym args ty -> SymFnCache sym -> SymFnCache sym
-insertSymFn s args ty fn = Map.alter upd s
-  where
-    upd Nothing = Just (MapF.singleton (Ctx.extend args ty) (SymFnWrapper fn))
-    upd (Just m) = Just (MapF.insert (Ctx.extend args ty) (SymFnWrapper fn) m)
-
-mkSymFn ::
-  forall sym args ret. (IsSymExprBuilder sym) =>
-  sym -> IORef (SymFnCache sym) ->
-  String -> Assignment BaseTypeRepr args -> BaseTypeRepr ret ->
-  IO (W.SymFn sym args ret)
-mkSymFn sym ref nm args ret =
-  do let s = W.safeSymbol nm
-     cache <- readIORef ref
-     case lookupSymFn s args ret cache of
-       Just fn -> return fn
-       Nothing ->
-         do fn <- W.freshTotalUninterpFn sym s args ret
-            writeIORef ref (insertSymFn s args ret fn cache)
-            return fn
-
-----------------------------------------------------------------------
--- Given a constant nm of (saw-core) type ty, construct an uninterpreted
--- constant with that type.
--- Importantly: The types of these uninterpreted values are *not*
--- limited to What4 BaseTypes or FirstOrderTypes.
-
-{- NOTE: How we generate uninterpreted functions
-
-Our strategy is to generate a separate uninterpreted function for each
-base type that is mentioned in the result of the original uninterpreted
-function.  If we need more than one value of a type, then we return an
-*array* of values, and the actual results of the original function are
-obtained by selecting elements of this array.
-
-For example, consider a function `f: [16] -> ([4][8],Bool)`.
-A call `f x` will be translated like this:
-
-  f_bv8:  [16] -> Array [2] [8]   // Uninterpreted
-  f_bool: [16] -> Bool            // Uninterpreted
-  w8s   = f_bv8 x                 // This has some auto-generated name
-  bools = f_bool x                // This has some auto-generated name
-
-Value for `f x`:
-  ([ w8s ! 0, w8s ! 1, w8s ! 2, w8s ! 3 ], bools)
--}
 
 
 
--- | Track how many uninterpreted results we need for each base type.
-newtype UnintCount (tc :: BaseType) = UnintCount Natural
-
--- | Uninterpreted function results for each base type.
-type UnintState sym = MapF BaseTypeRepr (Arr sym)
-
--- | Uninterpreted function results for a particular base type.
--- The length of the lists should match the @UnintCount tc@ for the type.
-newtype Arr sym tc = Arr [SymExpr sym tc]
-
--- | Generate a call to an uninterpreted function.
-parseUninterpreted ::
-  forall sym.
-  (IsSymExprBuilder sym) =>
-  sym                         {- ^ How to make symbolic expressions -} ->
-  IORef (SymFnCache sym)      {- ^ Cache of known uninterpreted functions -} ->
-  UnintApp (SymExpr sym)      {- ^ Name of function and its arguments -} ->
-  TValue (What4 sym)          {- ^ Type of the function's result -} ->
-  IO (SValue sym)             {- ^ The symbolic value for the call to the uninterpreted function -}
-parseUninterpreted sym ref app@(UnintApp nm args tys) ty =
-  do
-    un <- MapF.traverseWithKey mkUnint (countUninterpreted 1 MapF.empty ty)
-    evalStateT (parseUninterpreted' sym ref app ty) un
-  where
-  mkUnint :: BaseTypeRepr tc -> UnintCount tc -> IO (Arr sym tc)
-  mkUnint ret (UnintCount n)
-    | let lim = n - 1
-    , Just (Some w) <- someNat (width lim) =
-      case isZeroOrGT1 w of
-        Left Refl ->
-          do
-            fn <- mkSymFn sym ref (nm ++ suff ret) tys ret
-            el <- W.applySymFn sym fn args
-            pure (Arr [el])
-        Right LeqProof ->
-          do
-            fn   <- mkSymFn sym ref (nm ++ suff ret) tys (BaseArrayRepr (Ctx.singleton (BaseBVRepr w)) ret)
-            arr  <- W.applySymFn sym fn args
-            els <- forM [ 0 .. lim ] $ \i ->
-              do
-                ind <- W.bvLit sym w (BV.mkBV w (fromIntegral i))
-                W.arrayLookup sym arr (Ctx.singleton ind)
-            pure (Arr els)
-    | otherwise = panic "parseUninterpreted" ["`someNat` returned `Nothing` for `Natural`"]
-
-  -- | Compute the number of bits required to represent the given integer.
-  width :: Natural -> Natural
-  width x = go 0 x
-    where
-      go s 0 = s
-      go s n = let s' = s + 1 in s' `seq` go s' (n `shiftR` 1)
-
-  -- Type suffixes for uninterpreted functions of each base type.
-  suff :: BaseTypeRepr a -> String
-  suff t =
-    case t of
-      BaseBoolRepr -> "_bool"
-      BaseIntegerRepr -> "_int"
-      BaseRealRepr -> "_real"
-      BaseComplexRepr -> "_complex"
-      BaseBVRepr w -> "_bv" ++ show w
-      BaseFloatRepr (FloatingPointPrecisionRepr x y) -> "_float_" ++ show x ++ "_" ++ show y
-      BaseStringRepr w -> "_str" ++ strw
-        where strw = case w of
-                       Char8Repr -> "8"
-                       Char16Repr -> "16"
-                       UnicodeRepr -> "U"
-      BaseArrayRepr i e -> "_fun" ++ suff_asgn i ++ "_to" ++ suff e
-      BaseStructRepr flds -> "_struct" ++ suff_asgn flds ++ "_end"
-
-  suff_asgn i = intercalate "_and" (V.toList (Ctx.toVector i suff))
 
 
-
--- | Count how many uninterpreted symbols we need to represent a value
--- of the given type.  Note that this function should match exactly what
--- is needed by `parseUninterpreted'`.
-countUninterpreted ::
-  IsSymExprBuilder sym =>
-  Natural -> MapF BaseTypeRepr UnintCount -> TValue (What4 sym) -> MapF BaseTypeRepr UnintCount
-countUninterpreted scale count ty =
-  case ty of
-    VPiType {}      -> count
-    VBoolType       -> add BaseBoolRepr count
-    VIntType        -> add BaseIntegerRepr count
-    VIntModType {}  -> add BaseIntegerRepr count
-    VRationalType   -> add BaseIntegerRepr (add BaseIntegerRepr count)
-    VVecType n VBoolType ->
-      case somePosNat n of
-        Just (Some (PosNat w)) -> add (BaseBVRepr w) count
-        _                      -> count
-
-    VVecType n et -> if n == 0 then count else countUninterpreted (n * scale) count et
-
-    VArrayType ity ety
-      | Just (Some idx_repr) <- valueAsBaseType ity
-      , Just (Some elm_repr) <- valueAsBaseType ety
-      -> add (BaseArrayRepr (Ctx.Empty Ctx.:> idx_repr) elm_repr) count
-
-    VDataType (ModuleIdentifier "Prelude.UnitType") [] [] -> count
-
-    VDataType (ModuleIdentifier "Prelude.PairType") [TValue ty1, TValue ty2] [] ->
-      countUninterpreted scale (countUninterpreted scale count ty1) ty2
-
-    VDataType (ModuleIdentifier "Prelude.RecordType")
-      [VString _fname, TValue ty1, TValue ty2] [] ->
-      countUninterpreted scale (countUninterpreted scale count ty1) ty2
-
-    _ -> count
-  where
-  add ::
-    BaseTypeRepr tc ->
-    MapF BaseTypeRepr UnintCount ->
-    MapF BaseTypeRepr UnintCount
-  add bt =
-    MapF.insertWith
-      (\(UnintCount x) (UnintCount y) -> UnintCount (x + y))
-      bt (UnintCount scale)
-
--- Note that this doesn't really use IO, except for the `fail`.
-parseUninterpreted' ::
-  forall sym.
-  (IsSymExprBuilder sym) =>
-  sym ->
-  IORef (SymFnCache sym) ->
-  UnintApp (SymExpr sym) ->
-  TValue (What4 sym) ->
-  StateT (UnintState sym) IO (SValue sym)
-parseUninterpreted' sym ref app ty =
-  case ty of
-    VPiType _ body
-      -> pure $ VFun $ \x ->
-           do x' <- force x
-              app' <- applyUnintApp sym app x'
-              t2 <- applyPiBody body (ready x')
-              parseUninterpreted sym ref app' t2
-
-    VBoolType
-      -> VBool <$> mkUninterpreted BaseBoolRepr
-
-    VIntType
-      -> VInt  <$> mkUninterpreted BaseIntegerRepr
-
-    VIntModType n
-      -> VIntMod n <$> mkUninterpreted BaseIntegerRepr
-
-    VRationalType
-      -> do numer <- mkUninterpreted BaseIntegerRepr
-            -- TODO(#2433): Assert that the denominator is non-zero.
-            denom <- mkUninterpreted BaseIntegerRepr
-            pure $ VRational numer denom
-
-    VVecType n VBoolType ->
-      case somePosNat n of
-        Just (Some (PosNat w)) -> VWord . DBV <$> mkUninterpreted (BaseBVRepr w)
-        _                      -> pure (VWord ZBV)
-
-    VVecType n et ->
-      VVector <$> V.replicateM (fromIntegral n) (ready <$> parseUninterpreted' sym ref app et)
-
-    VArrayType ity ety
-      | Just (Some idx_repr) <- valueAsBaseType ity
-      , Just (Some elm_repr) <- valueAsBaseType ety
-      -> (VArray . SArray) <$>
-          mkUninterpreted (BaseArrayRepr (Ctx.Empty Ctx.:> idx_repr) elm_repr)
-
-    VDataType (ModuleIdentifier "Prelude.UnitType") [] []
-      -> pure vUnit
-
-    VDataType (ModuleIdentifier "Prelude.PairType") [TValue ty1, TValue ty2] []
-      -> do x1 <- parseUninterpreted' sym ref app ty1
-            x2 <- parseUninterpreted' sym ref app ty2
-            pure (vPair (ready x1) (ready x2))
-
-    VDataType (ModuleIdentifier "Prelude.EmptyType") [] []
-      -> pure vEmptyRecord
-    VDataType (ModuleIdentifier "Prelude.RecordType")
-      [VString _fname, TValue ty1, TValue ty2] []
-      -> do x1 <- parseUninterpreted' sym ref app ty1
-            x2 <- parseUninterpreted' sym ref app ty2
-            pure (vRecordValue (ready x1) (ready x2))
-
-    _ -> fail $ "could not create uninterpreted symbol of type " ++ show ty
-  where
-  mkUninterpreted :: BaseTypeRepr t -> StateT (UnintState sym) IO (SymExpr sym t)
-  mkUninterpreted tyr =
-    do mp <- get
-       case MapF.lookup tyr mp of
-         Just (Arr (x : xs)) -> put (MapF.insert tyr (Arr xs) mp) >> pure x
-         _ -> panic "mkUninterpreted" ["Not enough uninterpreted results"]
-
-
-
--- | A value of type @UnintApp f@ represents an uninterpreted function
--- with the given 'String' name, applied to a list of argument values
--- paired with a representation of their types. The context of
--- argument types is existentially quantified.
-data UnintApp f =
-  forall args. UnintApp String (Assignment f args) (Assignment BaseTypeRepr args)
-
--- | Extract the string from an 'UnintApp'.
-stringOfUnintApp :: UnintApp f -> String
-stringOfUnintApp (UnintApp s _ _) = s
-
--- | Make an 'UnintApp' with the given name and no arguments.
-mkUnintApp :: String -> UnintApp f
-mkUnintApp nm = UnintApp nm Ctx.empty Ctx.empty
-
--- | Add a suffix to the function name of an 'UnintApp'.
-suffixUnintApp :: String -> UnintApp f -> UnintApp f
-suffixUnintApp s (UnintApp nm args tys) = UnintApp (nm ++ s) args tys
-
--- | Extend an 'UnintApp' with an additional argument.
-extendUnintApp :: UnintApp f -> f ty -> BaseTypeRepr ty -> UnintApp f
-extendUnintApp (UnintApp nm xs tys) x ty =
-  UnintApp nm (Ctx.extend xs x) (Ctx.extend tys ty)
-
--- | Flatten an 'SValue' to a sequence of components, each of which is
--- a symbolic value of a base type (e.g. word or boolean), and add
--- them to the list of arguments of the given 'UnintApp'. If the
--- 'SValue' contains any values built from data constructors, then
--- encode them as suffixes on the function name of the 'UnintApp'.
-applyUnintApp ::
-  forall sym.
-  (W.IsExprBuilder sym) =>
-  sym ->
-  UnintApp (SymExpr sym) ->
-  SValue sym ->
-  IO (UnintApp (SymExpr sym))
-applyUnintApp sym app0 v =
-  case v of
-    VVector xv                -> foldM (applyUnintApp sym) app0 =<< traverse force xv
-    VBool sb                  -> return (extendUnintApp app0 sb BaseBoolRepr)
-    VInt si                   -> return (extendUnintApp app0 si BaseIntegerRepr)
-    VIntMod 0 si              -> return (extendUnintApp app0 si BaseIntegerRepr)
-    VIntMod n si              -> do n' <- W.intLit sym (toInteger n)
-                                    si' <- W.intMod sym si n'
-                                    return (extendUnintApp app0 si' BaseIntegerRepr)
-    VRational numer denom     -> do app1 <- applyUnintApp sym app0 (VInt numer)
-                                    app2 <- applyUnintApp sym app1 (VInt denom)
-                                    pure app2
-    VWord (DBV sw)            -> return (extendUnintApp app0 sw (W.exprType sw))
-    VArray (SArray sa)        -> return (extendUnintApp app0 sa (W.exprType sa))
-    VWord ZBV                 -> return app0
-    VCtorApp 0 _ []           -> return app0
-    VCtorApp 0 _ [x, y]       -> do app1 <- applyUnintApp sym app0 =<< force x
-                                    app2 <- applyUnintApp sym app1 =<< force y
-                                    return app2
-    VCtorApp n _ xs           -> foldM (applyUnintApp sym) app' =<< traverse force xs
-                                   where app' = suffixUnintApp ("_" ++ show n) app0
-    VNat n                    -> return (suffixUnintApp ("_" ++ show n) app0)
-    VBVToNat w v'             -> applyUnintApp sym app' v'
-                                   where app' = suffixUnintApp ("_" ++ show w) app0
-    TValue (suffixTValue -> Just s)
-                              -> return (suffixUnintApp s app0)
-    VFun {} ->
-      fail $
-      "Cannot create uninterpreted higher-order function " ++
-      show (stringOfUnintApp app0)
-    _ ->
-      fail $
-      "Cannot create uninterpreted function " ++
-      show (stringOfUnintApp app0) ++
-      " with argument " ++ show v
 
 
 ------------------------------------------------------------
@@ -1379,14 +1005,9 @@ argTypes v =
 
     loop _ = return []
 
---
--- Convert a saw-core type expression to a FirstOrder type expression
---
-vAsFirstOrderType :: IsSymExprBuilder sym => TValue (What4 sym) -> Maybe FirstOrderType
-vAsFirstOrderType v = asFirstOrderTypeTValue v
 
-valueAsBaseType :: IsSymExprBuilder sym => TValue (What4 sym) -> Maybe (Some W.BaseTypeRepr)
-valueAsBaseType v = fotToBaseType =<< vAsFirstOrderType v
+
+
 
 ------------------------------------------------------------------------------
 
@@ -1710,7 +1331,7 @@ w4EvalBasic sym st sc m addlPrims varCons ref unintSet t =
   do let variable tf (VarName ix nm) ty
            | Just v <- IntMap.lookup ix varCons = pure v
            | otherwise =
-           do trm <- ArgTermConst <$> scTermF sc tf
+           do trm <- scTermF sc tf
               parseUninterpretedSAW sym st sc ref trm
                  (mkUnintApp (Text.unpack nm ++ "_" ++ show ix)) ty
      let variable' tp vn ty = variable (Variable vn tp) vn ty
@@ -1723,336 +1344,6 @@ w4EvalBasic sym st sc m addlPrims varCons ref unintSet t =
      let mux = Prims.lazyMuxValue (prims sym)
      cfg <-
        Sim.evalGlobal' m (constMap sym `Map.union` addlPrims)
-       variable' uninterpreted (recursor sym) primHandler mux
+                        variable' uninterpreted (recursor sym) primHandler mux
      Sim.evalSharedTerm cfg t
 
--- | Given a constant nm of (saw-core) type ty, construct an
--- uninterpreted constant with that type. The 'Term' argument should
--- be an open term, which should have the designated return type when
--- the local variables have the corresponding types from the
--- 'Assignment'.
-parseUninterpretedSAW ::
-  forall n st fs.
-  B.ExprBuilder n st fs ->
-  SAWCoreState n ->
-  SharedContext ->
-  IORef (SymFnCache (B.ExprBuilder n st fs)) ->
-  ArgTerm {- ^ representation of function applied to arguments -} ->
-  UnintApp (SymExpr (B.ExprBuilder n st fs)) ->
-  TValue (What4 (B.ExprBuilder n st fs)) {- ^ return type -} ->
-  IO (SValue (B.ExprBuilder n st fs))
-parseUninterpretedSAW sym st sc ref trm app ty =
-  case ty of
-    VPiType t1 body
-      -> pure $ VFun $ \x ->
-           do x' <- force x
-              app' <- applyUnintApp sym app x'
-              arg <- mkArgTerm sc t1 x'
-              let trm' = ArgTermApply trm arg
-              t2 <- applyPiBody body (ready x')
-              parseUninterpretedSAW sym st sc ref trm' app' t2
-
-    VBoolType
-      -> VBool <$> mkUninterpretedSAW sym st sc ref trm app BaseBoolRepr
-
-    VIntType
-      -> VInt  <$> mkUninterpretedSAW sym st sc ref trm app BaseIntegerRepr
-
-    VIntModType n
-      -> VIntMod n <$> mkUninterpretedSAW sym st sc ref (ArgTermFromIntMod n trm) app BaseIntegerRepr
-
-    VRationalType
-      -> do numer <- mkUninterpretedSAW sym st sc ref trm (suffixUnintApp "_numer" app) BaseIntegerRepr
-            -- TODO(#2433): Assert that the denominator is non-zero.
-            denom <- mkUninterpretedSAW sym st sc ref trm (suffixUnintApp "_denom" app) BaseIntegerRepr
-            pure $ VRational numer denom
-
-    -- 0 width bitvector is a constant
-    VVecType 0 VBoolType
-      -> return $ VWord ZBV
-
-    VVecType n VBoolType
-      | Just (Some (PosNat w)) <- somePosNat n
-      -> (VWord . DBV) <$> mkUninterpretedSAW sym st sc ref trm app (BaseBVRepr w)
-
-    VVecType n ety | n >= 0
-      ->  do ety' <- termOfTValue sc ety
-             let mkElem i =
-                   do let trm' = ArgTermAt n ety' trm i
-                      let app' = suffixUnintApp ("_a" ++ show i) app
-                      parseUninterpretedSAW sym st sc ref trm' app' ety
-             xs <- traverse mkElem (genericTake n [0 ..])
-             return (VVector (V.fromList (map ready xs)))
-
-    VArrayType ity ety
-      | Just (Some idx_repr) <- valueAsBaseType ity
-      , Just (Some elm_repr) <- valueAsBaseType ety
-      -> (VArray . SArray) <$>
-          mkUninterpretedSAW sym st sc ref trm app (BaseArrayRepr (Ctx.Empty Ctx.:> idx_repr) elm_repr)
-
-    VDataType (ModuleIdentifier "Prelude.UnitType") [] []
-      -> pure vUnit
-
-    VDataType (ModuleIdentifier "Prelude.PairType") [TValue ty1, TValue ty2] []
-      -> do let trm1 = ArgTermPairLeft trm
-            let trm2 = ArgTermPairRight trm
-            x1 <- parseUninterpretedSAW sym st sc ref trm1 (suffixUnintApp "_L" app) ty1
-            x2 <- parseUninterpretedSAW sym st sc ref trm2 (suffixUnintApp "_R" app) ty2
-            return (vPair (ready x1) (ready x2))
-
-    VDataType (ModuleIdentifier "Prelude.EmptyType") [] []
-      -> pure vEmptyRecord
-    VDataType (ModuleIdentifier "Prelude.RecordType")
-      [VString fname, TValue ty1, TValue ty2] []
-      -> do let trm1 = ArgTermRecordSelect trm fname
-            let suffix = "_" ++ Text.unpack fname
-            x1 <- parseUninterpretedSAW sym st sc ref trm1 (suffixUnintApp suffix app) ty1
-            x2 <- parseUninterpretedSAW sym st sc ref trm app ty2
-            pure (vRecordValue (ready x1) (ready x2))
-
-    _ -> fail $ "could not create uninterpreted symbol of type " ++ show ty
-
-mkUninterpretedSAW ::
-  forall n st fs t.
-  B.ExprBuilder n st fs ->
-  SAWCoreState n ->
-  SharedContext ->
-  IORef (SymFnCache (B.ExprBuilder n st fs)) ->
-  ArgTerm ->
-  UnintApp (SymExpr (B.ExprBuilder n st fs)) ->
-  BaseTypeRepr t ->
-  IO (SymExpr (B.ExprBuilder n st fs) t)
-mkUninterpretedSAW sym st sc ref trm (UnintApp nm args tys) ret
-  | Just Refl <- testEquality Ctx.empty tys =
-    bindSAWTerm sym st ret =<< reconstructArgTerm trm sc []
-  | otherwise =
-  do fn <- mkSymFn sym ref nm tys ret
-     sawRegisterSymFunInterp st fn (reconstructArgTerm trm)
-     W.applySymFn sym fn args
-
--- | An 'ArgTerm' is a description of how to reassemble a saw-core
--- term from a list of the atomic components it is composed of.
-data ArgTerm
-  = ArgTermVar
-  | ArgTermBVZero -- ^ scBvNat 0 0
-  | ArgTermToIntMod Natural ArgTerm -- ^ toIntMod n x
-  | ArgTermFromIntMod Natural ArgTerm -- ^ fromIntMod n x
-  | ArgTermRational ArgTerm ArgTerm -- ^ numerator, denominator
-  | ArgTermVector Term [ArgTerm] -- ^ element type, elements
-  | ArgTermUnit
-  | ArgTermPair ArgTerm ArgTerm
-  | ArgTermEmpty
-  | ArgTermRecord FieldName ArgTerm ArgTerm
-  | ArgTermRecordSelect ArgTerm FieldName
-  | ArgTermConst Term
-  | ArgTermApply ArgTerm ArgTerm
-  | ArgTermAt Natural Term ArgTerm Natural
-    -- ^ length, element type, list, index
-  | ArgTermPairLeft ArgTerm
-  | ArgTermPairRight ArgTerm
-  | ArgTermBVToNat Natural ArgTerm
-
--- | Reassemble a saw-core term from an 'ArgTerm' and a list of parts.
--- The length of the list should be equal to the number of
--- 'ArgTermVar' constructors in the 'ArgTerm'.
-reconstructArgTerm :: ArgTerm -> SharedContext -> [Term] -> IO Term
-reconstructArgTerm atrm sc ts =
-  do (t, unused) <- parse atrm ts
-     unless (null unused) $ fail "reconstructArgTerm: too many function arguments"
-     return t
-  where
-    parse :: ArgTerm -> [Term] -> IO (Term, [Term])
-    parse at ts0 =
-      case at of
-        ArgTermVar ->
-          case ts0 of
-            x : ts1 -> return (x, ts1)
-            [] -> fail "reconstructArgTerm: too few function arguments"
-        ArgTermBVZero ->
-          do z <- scNat sc 0
-             x <- scBvNat sc z z
-             return (x, ts0)
-        ArgTermToIntMod n at1 ->
-          do n' <- scNat sc n
-             (x1, ts1) <- parse at1 ts0
-             x <- scToIntMod sc n' x1
-             pure (x, ts1)
-        ArgTermFromIntMod n at1 ->
-          do n' <- scNat sc n
-             (x1, ts1) <- parse at1 ts0
-             x <- scFromIntMod sc n' x1
-             pure (x, ts1)
-        ArgTermRational numer denom ->
-          do (numer', ts1) <- parse numer ts0
-             (denom', ts2) <- parse denom ts1
-             rat <- scRational sc numer' denom'
-             pure (rat, ts2)
-        ArgTermVector ty ats ->
-          do (xs, ts1) <- parseList ats ts0
-             x <- scVectorReduced sc ty xs
-             return (x, ts1)
-        ArgTermUnit ->
-          do x <- scUnitValue sc
-             return (x, ts0)
-        ArgTermPair at1 at2 ->
-          do (x1, ts1) <- parse at1 ts0
-             (x2, ts2) <- parse at2 ts1
-             x <- scPairValue sc x1 x2
-             return (x, ts2)
-        ArgTermEmpty ->
-          do x <- scRecordValue sc []
-             pure (x, ts0)
-        ArgTermRecord fname at1 at2 ->
-          do s <- scString sc fname
-             (x1, ts1) <- parse at1 ts0
-             (x2, ts2) <- parse at2 ts1
-             ty1 <- scTypeOf sc x1
-             ty2 <- scTypeOf sc x2
-             x <- scGlobalApply sc "Prelude.RecordValue" [s, ty1, ty2, x1, x2]
-             pure (x, ts2)
-        ArgTermRecordSelect at1 fname ->
-          do (x1, ts1) <- parse at1 ts0
-             x <- scRecordSelect sc x1 fname
-             pure (x, ts1)
-        ArgTermConst x ->
-          do return (x, ts0)
-        ArgTermApply at1 at2 ->
-          do (x1, ts1) <- parse at1 ts0
-             (x2, ts2) <- parse at2 ts1
-             x <- scApply sc x1 x2
-             return (x, ts2)
-        ArgTermAt n ty at1 i ->
-          do n' <- scNat sc n
-             (x1, ts1) <- parse at1 ts0
-             i' <- scNat sc i
-             x <- scAt sc n' ty x1 i'
-             return (x, ts1)
-        ArgTermPairLeft at1 ->
-          do (x1, ts1) <- parse at1 ts0
-             x <- scPairLeft sc x1
-             return (x, ts1)
-        ArgTermPairRight at1 ->
-          do (x1, ts1) <- parse at1 ts0
-             x <- scPairRight sc x1
-             return (x, ts1)
-        ArgTermBVToNat w at1 ->
-          do (x1, ts1) <- parse at1 ts0
-             x <- scBvToNat sc w x1
-             pure (x, ts1)
-
-    parseList :: [ArgTerm] -> [Term] -> IO ([Term], [Term])
-    parseList [] ts0 = return ([], ts0)
-    parseList (at : ats) ts0 =
-      do (x, ts1) <- parse at ts0
-         (xs, ts2) <- parseList ats ts1
-         return (x : xs, ts2)
-
--- | Given a type and value encoded as 'SValue's, construct an
--- 'ArgTerm' that builds a term of that type from local variables with
--- base types. The number of 'ArgTermVar' constructors should match
--- the number of arguments appended by 'applyUnintApp'.
-mkArgTerm :: forall sym. SharedContext -> TValue (What4 sym) -> SValue sym -> IO ArgTerm
-mkArgTerm sc ty val =
-  case (ty, val) of
-    (VBoolType, VBool _) -> return ArgTermVar
-    (VIntType, VInt _)   -> return ArgTermVar
-    (_, VWord ZBV)       -> return ArgTermBVZero     -- 0-width bitvector is a constant
-    (_, VWord (DBV _))   -> return ArgTermVar
-    (_, VArray{})        -> return ArgTermVar
-    (VIntModType n, VIntMod _ _) -> pure (ArgTermToIntMod n ArgTermVar)
-    (VRationalType, VRational numer denom) ->
-      do numer' <- mkArgTerm @sym sc VIntType (VInt numer)
-         denom' <- mkArgTerm @sym sc VIntType (VInt denom)
-         pure (ArgTermRational numer' denom')
-
-    (VVecType _ ety, VVector vv) ->
-      do vs <- traverse force (V.toList vv)
-         xs <- traverse (mkArgTerm sc ety) vs
-         ety' <- termOfTValue sc ety
-         return (ArgTermVector ety' xs)
-
-    (VDataType (ModuleIdentifier "Prelude.UnitType") [] [],
-     VCtorApp 0 _ [])
-                         -> return ArgTermUnit
-    (VDataType (ModuleIdentifier "Prelude.PairType") [TValue ty1, TValue ty2] [],
-     VCtorApp 0 _ [v1, v2]) ->
-      do x1 <- mkArgTerm sc ty1 =<< force v1
-         x2 <- mkArgTerm sc ty2 =<< force v2
-         return (ArgTermPair x1 x2)
-
-    (VDataType (ModuleIdentifier "Prelude.EmptyType") [] [],
-     VCtorApp 0 _ []) ->
-      pure ArgTermEmpty
-    (VDataType _nm [VString fname, TValue ty1, TValue ty2] [],
-     VCtorApp 0 _ [v1, v2]) ->
-      do x1 <- mkArgTerm sc ty1 =<< force v1
-         x2 <- mkArgTerm sc ty2 =<< force v2
-         pure (ArgTermRecord fname x1 x2)
-
-    (VDataType nmi ps _, VCtorApp n _ vv) ->
-      do mvi <- scResolveQualName sc (toQualName nmi)
-         vi <-
-           case mvi of
-             Just vi -> pure vi
-             Nothing -> panic "mkArgTerm" ["Data type name not found: " <> toAbsoluteName nmi]
-         mm <- scGetModuleMap sc
-         dt <-
-           case lookupVarIndexInMap vi mm of
-             Just (ResolvedDataType dt) -> pure dt
-             _ -> panic "mkArgTerm" ["Data type not found: " <> toAbsoluteName nmi]
-         let ctor = dtCtors dt !! n
-         ps' <- traverse (termOfSValue sc) ps
-         vv' <- traverse (termOfSValue sc <=< force) vv
-         x   <- scConstApply sc (ctorName ctor) (ps' ++ vv')
-         return (ArgTermConst x)
-
-    (_, TValue tval) ->
-      do x <- termOfTValue sc tval
-         pure (ArgTermConst x)
-
-    (_, VNat n) ->
-      do x <- scNat sc n
-         pure (ArgTermConst x)
-
-    (_, VBVToNat w v) ->
-      do let w' = fromIntegral w -- FIXME: make w :: Natural to avoid fromIntegral
-         x <- mkArgTerm sc (VVecType w' VBoolType) v
-         pure (ArgTermBVToNat w' x)
-
-    _ -> fail $ "could not create uninterpreted function argument of type " ++ show ty
-
-termOfTValue :: SharedContext -> TValue (What4 sym) -> IO Term
-termOfTValue sc val =
-  case val of
-    VBoolType -> scBoolType sc
-    VIntType -> scIntegerType sc
-    VRationalType -> scRationalType sc
-    VVecType n a ->
-      do n' <- scNat sc n
-         a' <- termOfTValue sc a
-         scVecType sc n' a'
-    VDataType (ModuleIdentifier "Prelude.UnitType") [] []
-      -> scUnitType sc
-    VDataType (ModuleIdentifier "Prelude.PairType") [TValue a, TValue b] []
-      -> do a' <- termOfTValue sc a
-            b' <- termOfTValue sc b
-            scPairType sc a' b'
-    VDataType (ModuleIdentifier "Prelude.EmptyType") [] []
-      -> scRecordType sc []
-    VDataType (ModuleIdentifier "Prelude.RecordType")
-      [VString fname, TValue a, TValue b] []
-      -> do fname' <- scString sc fname
-            a' <- termOfTValue sc a
-            b' <- termOfTValue sc b
-            scGlobalApply sc "Prelude.RecordType" [fname', a', b']
-    _ -> fail $ "termOfTValue: " ++ show val
-
-termOfSValue :: SharedContext -> SValue sym -> IO Term
-termOfSValue sc val =
-  case val of
-    VCtorApp 0 _ []
-      -> scUnitValue sc
-    VNat n
-      -> scNat sc n
-    TValue tv -> termOfTValue sc tv
-    _ -> fail $ "termOfSValue: " ++ show val